requirements.txt
openai==0.28
tiktoken
chromadb==0.4.4 
langchain==0.0.249
flask
GitPython
python-dotenv